More refactoring
While working on implementing the history endpoint, I realized that
passing querystring params to the endpoints is going to require some
refactoring for the base_request class.

As I started getting deeper into that class, I realized that the
`call_api` method is not really required as the `get` method is already
doing the necessary `if else` check.

I ended up cleaning a lot more than I initially anticipated but this
refactor feels like a step in the right direction and makes the codebase
more consistent

While running the tests locally, I had to rerecord the cassettes as the
get method adds an extra slash at the end of the URLs when the id passed
is nil.

As a result, I had to modify a couple of specs as elon's roadster has
moved further in its journey.