[Misc] Metaprogramming attempt
### Why?

Currently, we have to keep creating new classes for each endpoint that spacex api launches

These new classes are ***very** similar to each other

This is an attempt to avoid writing the main classes.

While this undeniably comes with the downsides of metaprogramming, I am wondering if we can alleviate some of that pain by continuing to write the individual specs and updating the readme file.